requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
schedule==1.2.0
python-dotenv==1.0.0
pandas==2.1.3
